+++ src/front-end/lists/static/js/custom/pages/home/controller-export-items-modal.js
@@ -1,0 +1,11 @@
+import { ExportItemsModal } from "./export-items-modal";
+
+
+
+/**********************************************************
+Action listeners for export items modal
+**********************************************************/
+export default function addExportItemsModalListeners() {
+    // show the spinner when the modal is closed
+    $(ExportItemsModal.Elements.MODAL).on('hidden.bs.modal', ExportItemsModal.showLoadingSection);
+}
